refactor(cli): decompose misc.rs junk drawer into focused command modules
Split the 1,615-line misc.rs catch-all into 10 dedicated command modules,
each owning a single CLI subcommand's implementation:

- benchmark.rs: provider latency benchmarking
- compress.rs: trajectory compression
- context.rs: project context file management
- info.rs: system overview and tool listing
- mcp.rs: MCP server management and stdio server
- model.rs: model listing, switching, browsing, and API verification
- pairing.rs: DM pairing authorization
- personality.rs: personality listing and preview
- plugins.rs: Lua plugin discovery and management
- toolset.rs: toolset distribution inspection

misc.rs is deleted. All references in lib.rs, init.rs, and slash.rs
updated to point to the new module paths. Public API unchanged.

Closes #275
